% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03-unitted.R
\name{as.data.frame.unitted}
\alias{as.data.frame.unitted}
\title{Construct a unitted element of a data.frame}
\usage{
\method{as.data.frame}{unitted}(x, ...)
}
\arguments{
\item{x}{A unitted object}

\item{...}{Other arguments passed to \code{as.data.frame()}}
}
\value{
A unitted data.frame element
}
\description{
\code{data.frames} are constructed by applying \code{as.data.frame()} to each
element. Thus, \code{as.data.frame.unitted()} is called whenever an argument
to \code{data.frame()} is unitted. \code{as.data.frame.unitted()} can handle
inputs of any subclass of \code{unitted}. Data.frame elements constructed
with as.data.frame.unitted continue to store their units as members of the
complete data.frame, although those units may not be visible until/unless the
data.frame is itself made unitted by a call to \code{u(mydf)}.
}

