% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-unitted.R
\name{as.data.frame.unitted_tbl_df}
\alias{as.data.frame.unitted_tbl_df}
\title{Override the default S3 for as.data.frame.tbl_df}
\usage{
\method{as.data.frame}{unitted_tbl_df}(x, ...)
}
\arguments{
\item{x}{The unitted_tbl_df to convert to a data.frame}

\item{...}{Additional arguments passed to as.data.frame}
}
\description{
Exporting this function seems to make it available for dispatch but still 
hidden. From here we could route back to the S4 dispatch system with 
".unitted_as.data.frame(object=x, ...)", but that seems like a waste of
processing time since we come here first.
}
\examples{
x <- as_data_frame(u(data.frame(x=u(1:3,"k"), y=u(3:5, "g"))))
as.data.frame(x)
}

