% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.list.unitted}
\alias{as.list.unitted}
\title{Convert unitted object to list}
\usage{
\method{as.list}{unitted}(x, ...)
}
\arguments{
\item{x}{object to be converted}

\item{...}{other arguments passed to class-specific methods}
}
\description{
Convert unitted object to list
}
