% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dplyr.R
\name{as_data_frame}
\alias{as_data_frame}
\alias{as_data_frame.default}
\alias{as_data_frame.unitted_data.frame}
\alias{as_data_frame.unitted_list}
\title{Implements dplyr::as_data_frame}
\usage{
as_data_frame(x)

\method{as_data_frame}{default}(x)

\method{as_data_frame}{unitted_list}(x)

\method{as_data_frame}{unitted_data.frame}(x)
}
\arguments{
\item{x}{A list, as in \code{\link[dplyr]{as_data_frame}}. Each element of
the list must have the same length.}
}
\description{
dplyr's \code{\link[dplyr]{as_data_frame}} serves other dplyr functions
including \code{\link[dplyr]{tbl_df}} and \code{\link[dplyr]{mutate}}.

Convert a data.frame or list to a tbl_df

Implements \code{\link[dplyr]{as_data_frame}} for unitted_data.frames

Implements \code{\link[dplyr]{as_data_frame}} for unitted_data.frames
}
\details{
This is not generic in dplyr but is made so here.
}
\examples{
as_data_frame(u(data.frame(x=1:3, y=3:5, z=c("aa", "bb", "cc")), c("X","Y","Z")))
as_data_frame(data.frame(x=1:3, y=3:5, z=c("aa", "bb", "cc")))
as_data_frame(list(x = 1:500, y = runif(500), z = 500:1))
as_data_frame(data.frame(x = 1:500, y = runif(500), z = 500:1))
as_data_frame(u(list(x=u(1:3,"X"), y=u(3:5,"Y"), z=u(c("aa", "bb", "cc"),"Z"))))
as_data_frame(u(data.frame(x=1:3, y=3:5, z=c("aa", "bb", "cc")), c("X","Y","Z")))
}

