% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{as_tibble.unitted_list}
\alias{as_tibble.unitted_data.frame}
\alias{as_tibble.unitted_list}
\alias{as_tibble.unitted_tbl_df}
\title{Convert a unitted_list to a unitted_tibble}
\usage{
\method{as_tibble}{unitted_list}(x)

\method{as_tibble}{unitted_data.frame}(x)

\method{as_tibble}{unitted_tbl_df}(x)
}
\arguments{
\item{x}{the unitted object to convert to a unitted_tbl_df}
}
\description{
Implements \code{\link[tibble]{as_tibble}} for unitted_lists

Implements \code{\link[tibble]{as_tibble}} for unitted_data.frames

Implements \code{\link[tibble]{as_tibble}} for unitted_tbl_dfs
}
\examples{
tibble::as_tibble(
 u(list(x=u(1:3,"X"), y=u(3:5,"Y"), z=u(c("aa", "bb", "cc"),"Z"))))
tibble::as_tibble(
 u(data.frame(x=1:3, y=3:5, z=c("aa", "bb", "cc")), c("X","Y","Z")))
tibble::as_tibble(
 u(data.frame(x=u(1:3,"X"), y=u(3:5,"Y"), z=u(c("aa", "bb", "cc"),"Z"))))
}

