% Generated by roxygen2 (4.0.1): do not edit by hand
\name{delimit_units}
\alias{delimit_units}
\title{Wrap a unit in delimiters if required}
\usage{
delimit_units(units, delimiter, rule)
}
\arguments{
\item{units}{Character string[s] each containing a single unit}

\item{delimiter}{A single-character string indicating the delimiter to be
used}

\item{rule}{Character string indicating the delimitation rule to follow.}
}
\value{
A character vector of the same length as \code{units}, with each
  string delimited if required by \code{rule}.
}
\description{
A helper to merge_units that wraps the unit in the delimiter, or not,
according to \code{rule}.
}
\seealso{
Other unit-merging functions: \code{\link{merge_units}}
}

