% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{.str}
\alias{.str}
\title{importFrom utils str}
\usage{
.str(object, ...)
}
\arguments{
\item{object}{the object to describe}

\item{...}{other args for future development}
}
\description{
importFrom utils str
}

