% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_unitbundles}
\alias{get_unitbundles}
\title{The internal, reasonably efficient method for acquiring an object's units in
the internal representation}
\usage{
get_unitbundles(object, ...)
}
\arguments{
\item{object}{The object whose units should be returned}
}
\value{
A unitbundle or list of unitbundles, each representing one set of
  units
}
\description{
get_unitbundles() of a non-unitted class returns NA.
}
\details{
get_unitbundles() of a unitted vector, array, or matrix class returns the single
unitbundle associated with that object.

get_unitbundles() of a data.frame or unitted_data.frame returns a named list of
unitbundles (or NAs), one per data.frame column.
}

