% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{get_units,unitted-method}
\alias{get_units,unitted-method}
\title{The exported, client-side method for acquiring an object's units in string representation.}
\usage{
\S4method{get_units}{unitted}(object, ...)
}
\arguments{
\item{y}{The object whose units should be returned}

\item{object}{The object from which to retrieve units}

\item{...}{Other arguments passed on to the type-specific implementations}
}
\value{
A string representing one set of units
}
\description{
The exported, client-side method for acquiring an object's units in string representation.
}

