% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/02-unitbundle.R, R/inspect.R
\docType{methods}
\name{get_units}
\alias{get_units}
\alias{get_units,ANY-method}
\alias{get_units,data.frame-method}
\alias{get_units,list-method}
\alias{get_units,unitbundle-method}
\alias{get_units,unitted-method}
\alias{get_units,unitted_list-method}
\title{Gets the units as a string or strings.}
\usage{
get_units(object, ...)

\S4method{get_units}{ANY}(object, ...)

\S4method{get_units}{unitbundle}(object, delimiter = "|",
  rule = c("disambiguate", "never", "always"), separator = " ", ...)

\S4method{get_units}{unitted}(object, ...)

\S4method{get_units}{unitted_list}(object, recursive = FALSE, ...)

\S4method{get_units}{list}(object, recursive = TRUE, ...)

\S4method{get_units}{data.frame}(object, recursive = TRUE, ...)
}
\arguments{
\item{object}{The object from which to retrieve units}

\item{...}{Other arguments passed on to the type-specific implementations}

\item{delimiter}{A single-character string designating the delimiter that
should surround those units that are to be delimited according to
\code{rule}}

\item{rule}{character string indicating the rule by which each unit within a
unitdf will be delimited or not. The default, "disambiguate", wraps only
those units that contain spaces or "^" characters. "never" wraps none, and
"always" wraps all.}

\item{separator}{character indicating a split among pieces of a unit, e.g.,
"mg L^-1" is separated by " ".}

\item{recursive}{logical. If \code{object} is a list or data.frame, should
the units attached to each element be returned?}
}
\value{
A string representing one set of units
}
\description{
Units can be acquired from objects of all types; types without units will
return NA.

The exported, client-side method for acquiring an object's units in string
representation.
}
\examples{
get_units(5)
get_units(unitbundle("king kong"))
}

