% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inspect.R
\name{is.unitted}
\alias{is.unitted}
\title{Determine whether an object is unitted}
\usage{
is.unitted(x)
}
\arguments{
\item{x}{The potentially unitted object}
}
\value{
Logical. TRUE if x is unitted, FALSE otherwise
}
\description{
Determine whether an object is unitted
}

