% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_join.R
\name{join}
\alias{join}
\alias{inner_join.unitted_data.frame}
\alias{left_join.unitted_data.frame}
\alias{right_join.unitted_data.frame}
\alias{full_join.unitted_data.frame}
\alias{semi_join.unitted_data.frame}
\alias{anti_join.unitted_data.frame}
\alias{inner_join.unitted_tbl_df}
\alias{left_join.unitted_tbl_df}
\alias{right_join.unitted_tbl_df}
\alias{full_join.unitted_tbl_df}
\alias{semi_join.unitted_tbl_df}
\alias{anti_join.unitted_tbl_df}
\title{Join two unitted data.frames}
\usage{
\method{inner_join}{unitted_data.frame}(x, y, by = NULL, copy = FALSE, ...)

\method{left_join}{unitted_data.frame}(x, y, by = NULL, copy = FALSE, ...)

\method{right_join}{unitted_data.frame}(x, y, by = NULL, copy = FALSE, ...)

\method{full_join}{unitted_data.frame}(x, y, by = NULL, copy = FALSE, ...)

\method{semi_join}{unitted_data.frame}(x, y, by = NULL, copy = FALSE, ...)

\method{anti_join}{unitted_data.frame}(x, y, by = NULL, copy = FALSE, ...)

\method{inner_join}{unitted_tbl_df}(x, y, by = NULL, copy = FALSE, ...)

\method{left_join}{unitted_tbl_df}(x, y, by = NULL, copy = FALSE, ...)

\method{right_join}{unitted_tbl_df}(x, y, by = NULL, copy = FALSE, ...)

\method{full_join}{unitted_tbl_df}(x, y, by = NULL, copy = FALSE, ...)

\method{semi_join}{unitted_tbl_df}(x, y, by = NULL, copy = FALSE, ...)

\method{anti_join}{unitted_tbl_df}(x, y, by = NULL, copy = FALSE, ...)
}
\arguments{
\item{x}{left-hand-side table to join}

\item{y}{right-hand-side table to join}

\item{by}{A join specification created with \code{\link[dplyr:join_by]{join_by()}}, or a character
vector of variables to join by.

If \code{NULL}, the default, \verb{*_join()} will perform a natural join, using all
variables in common across \code{x} and \code{y}. A message lists the variables so
that you can check they're correct; suppress the message by supplying \code{by}
explicitly.

To join on different variables between \code{x} and \code{y}, use a \code{\link[dplyr:join_by]{join_by()}}
specification. For example, \code{join_by(a == b)} will match \code{x$a} to \code{y$b}.

To join by multiple variables, use a \code{\link[dplyr:join_by]{join_by()}} specification with
multiple expressions. For example, \code{join_by(a == b, c == d)} will match
\code{x$a} to \code{y$b} and \code{x$c} to \code{y$d}. If the column names are the same between
\code{x} and \code{y}, you can shorten this by listing only the variable names, like
\code{join_by(a, c)}.

\code{\link[dplyr:join_by]{join_by()}} can also be used to perform inequality, rolling, and overlap
joins. See the documentation at \link[dplyr:join_by]{?join_by} for details on
these types of joins.

For simple equality joins, you can alternatively specify a character vector
of variable names to join by. For example, \code{by = c("a", "b")} joins \code{x$a}
to \code{y$a} and \code{x$b} to \code{y$b}. If variable names differ between \code{x} and \code{y},
use a named character vector like \code{by = c("x_a" = "y_a", "x_b" = "y_b")}.

To perform a cross-join, generating all combinations of \code{x} and \code{y}, see
\code{\link[dplyr:cross_join]{cross_join()}}.}

\item{copy}{If \code{x} and \code{y} are not from the same data source,
and \code{copy} is \code{TRUE}, then \code{y} will be copied into the
same src as \code{x}.  This allows you to join tables across srcs, but
it is a potentially expensive operation so you must opt into it.}

\item{...}{Other parameters passed onto methods.}
}
\description{
Uses the corresponding dplyr join functions.
}
