% Generated by roxygen2 (4.0.1): do not edit by hand
\name{merge_units}
\alias{merge_units}
\title{Turn a units data.frame into the character representation of the units}
\usage{
merge_units(unitdfs, delimiter = "|", rule = c("disambiguate", "never",
  "always"), separator = " ")
}
\arguments{
\item{unitdfs}{list of unit data.frames, such as those returned by a call to
parse_units}

\item{delimiter}{A single-character string designating the delimiter that
should surround those units that are to be delimited according to
\code{rule}}

\item{rule}{character string indicating the rule by which each
unit within a unitdf will be delimited or not. The default, "disambiguate",
wraps only those units that contain spaces or "^" characters. "never" wraps
none, and "always" wraps all.}
}
\value{
list of character strings, one for each data.frame passed in
}
\description{
Turn a units data.frame into the character representation of the units
}
\seealso{
Other unit-merging functions: \code{\link{delimit_units}}
}

