% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{mutate}
\alias{mutate}
\alias{mutate.unitted_data.frame}
\alias{mutate.unitted_tbl_df}
\title{Implements dplyr::mutate and dplyr::mutate_ for unitted_data.frames}
\usage{
\method{mutate}{unitted_data.frame}(.data, ...)

\method{mutate}{unitted_tbl_df}(.data, ...)
}
\arguments{
\item{.data}{a unitted_data.frame}

\item{...}{standard dots, as in \code{\link[dplyr]{select_}}}
}
\value{
a unitted_data.frame after the \code{\link[dplyr]{select_}} operation
}
\description{
Implements dplyr::mutate and dplyr::mutate_ for unitted_data.frames

Implements dplyr::mutate and dplyr::mutate_ for unitted_tbl_dfs
}
\examples{
df <- u(data.frame(x=1:3, y=3:5), c("X","Y"))
dplyr::mutate(df, z=LETTERS[y], k=x*y)
dplyr::mutate(df, z=u(LETTERS[y],'LL'), k=x*y)
dtb <- tibble::as_tibble(u(data.frame(x=1:3, y=3:5), c("X","X")))
dplyr::mutate(dtb, z=LETTERS[y], k=x+y)
dplyr::mutate(dtb, z=u(LETTERS[y],'LL'), k=x+y)
}
