% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{mutate_.unitted_tbl_df}
\alias{mutate_.unitted_tbl_df}
\title{Implements dplyr::mutate and dplyr::mutate_ for unitted_tbl_dfs}
\usage{
\method{mutate_}{unitted_tbl_df}(.data, ..., .dots)
}
\arguments{
\item{.data}{a unitted_tbl_df}

\item{...}{standard dots, as in \code{\link[dplyr]{select_}}}

\item{.dots}{nonstandard dots, as in \code{\link[dplyr]{select_}}}
}
\value{
a unitted_tbl_df after the \code{\link[dplyr]{select_}} operation
}
\description{
Implements dplyr::mutate and dplyr::mutate_ for unitted_tbl_dfs
}
\examples{
dtb <- tibble::as_tibble(u(data.frame(x=1:3, y=3:5), c("X","X")))
dplyr::mutate(dtb, z=LETTERS[y], k=x+y)
}
