% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{package-unitted}
\alias{package-unitted}
\alias{package-unitted-package}
\title{Rigorous unit tracking during data manipulation}
\description{
\code{unitted} helps you track the units of data values through all common
data manipulation activities - addition, multiplication, exponentiation,
subsetting, and so on.
}

