% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.unitted}
\alias{print.unitted}
\title{Print a unitted object}
\usage{
\method{print}{unitted}(x, ...)
}
\arguments{
\item{x}{The unitted object}

\item{value}{The value or values to insert in the specified cell[s]}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An invisible unitted vector
}
\description{
For many data types, \code{print.unitted} includes the units in its output.
For data.frames, for example, it adds a second header row to display the
units of each column.
}

