% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{select}
\alias{select}
\alias{select.unitted_data.frame}
\alias{select.unitted_tbl_df}
\alias{rename.unitted_data.frame}
\alias{rename.unitted_tbl_df}
\title{Select/rename variables by name for unitted data.frames & tibbles}
\usage{
\method{select}{unitted_data.frame}(.data, ...)

\method{select}{unitted_tbl_df}(.data, ...)

\method{rename}{unitted_data.frame}(.data, ...)

\method{rename}{unitted_tbl_df}(.data, ...)
}
\arguments{
\item{.data}{a unitted_data.frame}

\item{...}{standard dots, as in \code{\link[dplyr]{rename_}}}
}
\value{
a unitted_data.frame after the \code{\link[dplyr]{rename_}} operation

a unitted_data.frame after the \code{\link[dplyr]{rename_}} operation
}
\description{
Select/rename variables by name for unitted data.frames & tibbles

Implements dplyr::select and dplyr::select_ for unitted_data.frames

Implements dplyr::select and dplyr::select_ for unitted_tbl_dfs

Implements dplyr::rename and dplyr::rename_ for unitted_data.frames

Implements dplyr::rename and dplyr::rename_ for unitted_tbl_dfs
}
\examples{
dplyr::select(
 u(data.frame(x=1:3, y=3:5, z=c("aa", "bb", "cc")), c("X","Y","Z")), 
 a=y, x)
dplyr::select(
 tibble::as_tibble(u(data.frame(x=1:3, y=3:5, z=c("aa", "bb", "cc")), c("X","Y","Z"))), 
 a=y, x)
df <- u(data.frame(x=1:3, y=3:5, z=c("aa", "bb", "cc")), c("X","Y","Z"))
dplyr::rename(df, a=y, beta=x)
df <- u(data.frame(x=1:3, y=3:5, z=c("aa", "bb", "cc")), c("X","Y","Z"))
dplyr::rename(tibble::as_tibble(df), a=y, beta=x)
}
