% Generated by roxygen2 (4.0.1): do not edit by hand
\name{separate_units}
\alias{separate_units}
\title{Separate a unitbundle into a data frame of units and powers}
\usage{
separate_units(unitbdl)
}
\arguments{
\item{unitbds}{a unitbundle}
}
\value{
a data.frame with columns for unit and power
}
\description{
Separate a unitbundle into a data frame of units and powers
}

