% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,unitted-method}
\alias{show,unitted-method}
\title{Show a unitted object}
\usage{
\S4method{show}{unitted}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
Displays a unitted object, including its units and data type whenever possible
}

