% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01-parse.R
\name{sort_units}
\alias{sort_units}
\title{Sort the units data.frame by numerator/denominator and then alphabetically by
the unit name}
\usage{
sort_units(unitdfs)
}
\arguments{
\item{unitdfs}{list of unit data.frames (having columns Unit and Power)}
}
\value{
list of unit data.frames, each one sorted by num/denom and by unit
  name
}
\description{
Sort the units data.frame by numerator/denominator and then alphabetically by
the unit name
}
\seealso{
Other unit-parsing functions: \code{\link{parse_units}};
  \code{\link{prepare_delimiter}};
  \code{\link{simplify_units}}
}

