% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dplyr.R
\name{tbl_df}
\alias{tbl_df}
\title{Implements dplyr::tbl_df}
\usage{
tbl_df(data)
}
\arguments{
\item{data}{A list, as in \code{\link[dplyr]{as_data_frame}}. Each element of
the list must have the same length.}
}
\description{
This is not generic in dplyr but is made so here.
}
\examples{
tbl_df(u(data.frame(x=1:3, y=3:5, z=c("aa", "bb", "cc")), c("X","Y","Z")))
tbl_df(data.frame(x=1:3, y=3:5, z=c("aa", "bb", "cc")))
}

