% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-unitbundle.R
\docType{class}
\name{unitbundle-class}
\alias{Constructor}
\alias{unitbundle}
\alias{unitbundle-class}
\title{Class unitbundle}
\usage{
unitbundle(units, delimiter = "|", ..., unitdf, unitstr, unitbdl)
}
\arguments{
\item{units}{A data.frame (unitdf method), character (parse string method), 
or unitbundle (replica method). Although the code is cleaner, it is 
slightly less efficient to use this units argument than to explicitly 
specify unitdf, unitstr, or unitbdl.}

\item{delimiter}{A single-character string, relevant only when \code{units} 
or \code{unitstr} is a character string, used in parsing the units string 
into a bundle of individual units. For example, the string "tree house^2" 
would be parsed as two units, "tree" and "house"^2, but the default 
delimiter of "|" causes the string "|tree house|^2" to be parsed as one 
squared unit, "tree house"^2.}

\item{...}{Other arguments passed to initialize}

\item{unitdf}{A data.frame with columns Unit=character, Power=numeric, 
specifying each unit-power pair in a separate row.}

\item{unitstr}{A single character string containing one or more units to be 
parsed and bundled, e.g. \code{"kg ha^-1 yr^-1"}}

\item{unitbdl}{A unitbundle. If this is supplied, the new unitbundle is 
simply a replica of unitbdl.}
}
\value{
An initialized \code{unitbundle}
}
\description{
A unitbundle is an object containing information about a single set of units 
- for example, \eqn{kg*m^2*s^-2}{kg m^2 s^-2} or \eqn{mi*hr^-1}{mi hr^-1}. 
This information is stored in a way that permits efficient units conversion 
and arithmetic.

There are several ways to create a unitbundle, all of which use the standard 
constructor \code{unitbundle(...)}. Exactly one or zero arguments from 
[\code{units}, \code{unitstr}, \code{unitdf}, or \code{unitbdl}] may be 
specified; the choice or argument determines which construction method is 
used. If the \code{units} argument is specified, with or without the 'units' 
label, then the class of that argument determines whether it is interpreted 
as a \code{unitdf} argument (if it's a data.frame), a \code{unitstr} argument
(if it's a character), or a \code{unitbdl} argument (if it's a unitbundle).
}
\details{
The basic content of a unitbundle is a collection of units each raised to a 
power (see Definitions below). For example, \code{"kg ha^-1"} has two units, 
\code{"kg"}, and \code{"ha"}, raised to the powers \code{1} and \code{-1}, 
respectively. In creating a unitbundle from a character string, the caret 
symbol (\code{^}) is used to indicate the separation between the unit and its
power, and a space usually indicates the separation between two unit-power 
pairs.

Character strings are a potentially ambiguous way to describe units 
information; in particular, a unit might itself contain a space or \code{^}. 
To specify the contents of a unitbundle without ambiguity, a \code{delimiter}
may be used to indicate where the text part of a unit begins and ends. For 
example, \code{"g soil^-1"} will be incorrectly parsed into \code{"g"} and 
\code{"soil^-1"}, but this can be avoided by wrapping the two-word unit in 
the delimiter: when \code{delimiter="|"}, then \code{"|g soil|^-1"} will be
parsed such that \code{"g soil"} is a single unit.
}
\section{Definitions}{

  
  \describe{
  
  \item{unit}{A single property or thing: \eqn{kg}, \eqn{tree*house}{tree
  house}, \eqn{"kgC_organic"}{kgC_organic}}
  
  \item{power}{The power to which a unit is raised; assumed to be 1 if 
  unspecified. \eqn{1}, \eqn{0.53}, and \eqn{1/2} are powers in unit-power 
  pairs such as \code{"apples^1"} or \code{"apples"}, 
  \code{"frequency^0.53"}, and \code{"acre^1/2"}}
  
  \item{unit string}{A combination of units and powers used to describe data:
  \code{"kgC ha^-1 yr^-1"}}
  
  \item{vector of unit strings}{\code{c("kgC ha^-1 yr^-1", "PgN", "ft^3 
  s^-1")}}
  
  }
}
\examples{
unitbundle()
unitbundle("kg ha^-1 yr^-1")
unitbundle(units="kg ha^-1 yr^-1")
unitbundle(unitstr="kg ha^-1 yr^-1")
unitbundle(unitdf=data.frame(Unit=c("kg","ha","yr"), Power=c(1,-1,-1)))
unitbundle(unitbdl=unitbundle("kg ha^-1 yr^-1"))
}
\seealso{
\code{\linkS4class{unitted}} for data with unitbundles attached; 
  \code{\link{unitbundle_Ops}} for arithmetic and other operations on
  unitbundles; \code{\link{get_units}} for conversion to character formats
}

