% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-unitbundle.R
\docType{methods}
\name{Arith,unitbundle,unitbundle-method}
\alias{Arith,unitbundle,unitbundle-method}
\title{Operations on unitbundles}
\usage{
\S4method{Arith}{unitbundle,unitbundle}(e1, e2)
}
\arguments{
\item{e1}{the first unitbundle}

\item{e2}{the second unitbundle}
}
\description{
When data with units are combined by arithmetic operations, the units ought 
to be propagated through according to standard rules. Data that are 
\code{unitted} obey these rules. To make this possible, the units themselves 
(as \code{unitbundle} objects) also obey rules relevant to units propagation,
but be warned - the rules are slightly different for \code{unitbundle} 
objects than they are for \code{unitted} objects. See \strong{Details | 
Arithmetic on unitbundles}.
}
\details{
The rules for operations on unitbundles are described in the following
subsections.

\subsection{Arithmetic on unitbundles}{

\describe{

\item{+e1, -e1, !e1}{Returns the original units of e1, which are unchanged by
these unary operations.}

\item{e1+e2, e1-e2}{e1 and e2 are required to contain identical units (an 
error is thrown otherwise); the units of e1 (or, equivalently, e2) are 
returned unaltered.}

\item{e1*e2}{Returns a new unitbundle containing the union (product) of all 
units in e1 and e2.}

\item{e1/e2}{Returns a new unitbundle with the units of e1 in the numerator 
and the units of e2 in the denominator.}

\item{e1^e2}{Requires that e2 contain no units; the new units must be 
calculated with respect to the data attached to e2 but cannot be calculated 
here because the data are unavailable to unitbundles themselves. The units of
e1 are returned.}

\item{e1\%\%e2}{Returns the units of e1 regardless of the units of e2.}

\item{e1\%/\%e2}{Returns the same units as e1/e2.}

}

}
}
\seealso{
\code{\link{unitbundle}} for the \code{unitbundle} class; 
  \code{\linkS4class{unitted}} for data with unitbundles attached
}

