% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-unitbundle.R
\docType{methods}
\name{Compare,unitbundle,unitbundle-method}
\alias{Compare,unitbundle,unitbundle-method}
\title{Comparisons of unitbundles}
\usage{
\S4method{Compare}{unitbundle,unitbundle}(e1, e2)
}
\arguments{
\item{e1}{the first unitbundle}

\item{e2}{the second unitbundle}
}
\description{
When data with units are compared, the units ought 
to be considered according to standard rules. Data that are 
\code{unitted} obey these rules. To make this possible, the units themselves 
(as \code{unitbundle} objects) also obey rules relevant to comparison,
but be warned - the rules are slightly different for \code{unitbundle} 
objects than they are for \code{unitted} objects. See \strong{Details |
Comparisons of unitbundles}.
}
\details{
\subsection{Comparisons of unitbundles}{

\describe{

\item{e1==e2}{True if e1 and e2 are identical.}

\item{e1!=e2}{True if e1 and e2 are anything but identical.}

\item{>, <, >=, <=}{Undefined for unitbundles.}

}

}
}
