% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/02-unitbundle.R
\docType{methods}
\name{Logic,unitbundle,unitbundle-method}
\alias{Logic,unitbundle,unitbundle-method}
\title{Logical operations on unitbundles}
\usage{
\S4method{Logic}{unitbundle,unitbundle}(e1, e2)
}
\arguments{
\item{e1}{the first unitbundle}

\item{e2}{the second unitbundle}
}
\description{
It is not clear what it would mean to perform logical operations on
unitbundles; for this reason, methods in the \code{Logic} group throw errors
when applied to unitbundles.
}

