% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-unitted.R
\docType{package}
\name{unitted-package}
\alias{unitted-package}
\title{unitted: Track Units as you Manipulate Data}
\description{
Attaches units to data.frames, vectors, etc. and maintains the correct units when you add, multiply, exponentiate, etc. those values.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/appling/unitted}
}

}
\keyword{internal}
