% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{unitted_assign}
\alias{$<-}
\alias{$<-,unitted_data.frame-method}
\alias{Extract}
\alias{[<-}
\alias{[<-,unitted_data.frame,ANY,ANY-method}
\alias{[<-,unitted_data.frame,ANY,missing-method}
\alias{[<-,unitted_data.frame,missing,ANY-method}
\alias{[<-,unitted_data.frame,missing,missing-method}
\alias{unitted_assign}
\title{Replace parts of a unitted vector, array, or data.frame}
\usage{
\S4method{[}{unitted,ANY,missing}(x, i, j, ...) <- value

\S4method{$}{unitted_data.frame}(x, name) <- value

\S4method{[}{unitted_data.frame,ANY,ANY}(x, i, j, ...) <- value

\S4method{[}{unitted_data.frame,ANY,missing}(x, i, j, ...) <- value

\S4method{[}{unitted_data.frame,missing,ANY}(x, i, j, ...) <- value

\S4method{[}{unitted_data.frame,missing,missing}(x, i, j, ...) <- value
}
\arguments{
\item{x}{The unitted data.frame, vector, etc. to be partially replaced}

\item{i}{(optional) The first replacement index}

\item{j}{(optional) The second replacement index}

\item{...}{Other arguments passed to replacement functions}

\item{value}{The value(s) with which to replace the indexed elements of x}

\item{name}{The name of the column to be replaced}
}
\value{
\code{x} with its indexed elements replaced by \code{value}
}
\description{
Replace parts of a unitted vector, array, or data.frame
}
\seealso{
\code{\link{unitted_access}} for assignment to parts of objects;
  \code{\linkS4class{unitted}} for definition of the unitted class
}

