% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{unitted_bind}
\alias{unitted_bind}
\alias{rbind.unitted}
\alias{unitted_rbind}
\alias{rbind}
\alias{bind}
\alias{cbind.unitted}
\alias{unitted_cbind}
\alias{cbind}
\title{Bind unitted objects by row or column}
\usage{
\method{rbind}{unitted}(...)

\method{cbind}{unitted}(...)
}
\arguments{
\item{...}{unitted vectors, matrices, or data.frames}
}
\description{
Combines unitted objects as if they were not unitted, but enforces unit
consistency across the objects to be combined.
}
\details{
\code{deparse.level} is an integer controlling the construction of labels, as in 
  the default rbind and cbind methods. Only deparse.level = 0 and 1 are 
  available for unitted rbind/cbind calls.
}
\seealso{
Other unitted object manipulation: \code{\link{unitted_c}},
  \code{\link{unitted_merge}}, \code{\link{unitted_rep}}
}
