% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-unitted.R
\docType{class}
\name{unitted_data.frame-class}
\alias{unitted_data.frame-class}
\title{A data.frame with units attached to each column}
\description{
Data frame columns can be extracted with units and cells can be modified with
units checking.
}

