% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_join.R
\name{unitted_half_join.unitted_data.frame}
\alias{unitted_half_join.unitted_data.frame}
\title{Internal function to join two unitted data.frames}
\usage{
unitted_half_join.unitted_data.frame(ux, uy, join_fun = inner_join, ...)
}
\description{
Appropriate to semi_join and anti_join, both of which retain the columns and
units of the first data.frame
}
\keyword{internal}
