% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_join.R
\name{unitted_join.unitted_data.frame}
\alias{unitted_join.unitted_data.frame}
\title{Internal function to join two unitted data.frames}
\usage{
unitted_join.unitted_data.frame(ux, uy, join_fun = inner_join, ...)
}
\description{
Appropriate to full, inner, left, and right joins, where the units are those
resulting from a separate join on the units alone
}
\keyword{internal}

