% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine.R
\docType{methods}
\name{unitted_merge}
\alias{merge,ANY,unitted-method}
\alias{merge,unitted,ANY-method}
\alias{unitted_merge}
\title{Merge unitted data.frames by one or more common columns}
\usage{
\S4method{merge}{unitted_data.frame,unitted_data.frame}(x, y, ...)

\S4method{merge}{unitted,ANY}(x, y, ...)

\S4method{merge}{ANY,unitted}(x, y, ...)
}
\description{
Merges unitted data.frames, ensuring units compatibility among the common
columns
}
\seealso{
\code{base::\link{merge}}

Other unitted object manipulation: \code{\link{bind}},
  \code{\link{cbind}}, \code{\link{cbind.unitted}},
  \code{\link{rbind}}, \code{\link{rbind.unitted}},
  \code{\link{unitted_bind}}, \code{\link{unitted_cbind}},
  \code{\link{unitted_rbind}}; \code{\link{c.unitted}},
  \code{\link{unitted_c}}; \code{\link{rep.unitted}},
  \code{\link{rep.unitted_data.frame}},
  \code{\link{unitted_rep}}
}

