% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dplyr.R
\name{unitted_rename}
\alias{unitted_rename}
\title{Internal function for renaming unitted objects}
\usage{
unitted_rename(.data, .dots)
}
\description{
Internal function for renaming unitted objects
}
\keyword{internal}

