% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{unitted_rep}
\alias{unitted_rep}
\alias{rep.unitted}
\alias{rep.unitted_data.frame}
\title{Replicate elements, maintaining the original units}
\usage{
\method{rep}{unitted}(x, ...)

\method{rep}{unitted_data.frame}(x, ...)
}
\arguments{
\item{x}{object to replicate}

\item{...}{other objects passed to \code{rep}}
}
\description{
Wrapper for non-unitted rep() methods.
}
\seealso{
Other unitted object manipulation: \code{\link{unitted_bind}},
  \code{\link{unitted_c}}, \code{\link{unitted_merge}}
}
