% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{unitted_select}
\alias{unitted_select}
\title{Internal function for selecting from unitted objects}
\usage{
unitted_select(.data, .dots)
}
\description{
Internal function for selecting from unitted objects
}
\keyword{internal}

