% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/display.R
\name{unitted_str}
\alias{str.unitted}
\alias{str.unitted_NULL}
\alias{str.unitted_data.frame}
\alias{unitted_str}
\title{Return the structure of a unitted object.}
\usage{
\method{str}{unitted}(object, ...)

\method{str}{unitted_data.frame}(object, ...)

\method{str}{unitted_NULL}(object, ...)
}
\arguments{
\item{object}{The object whose structure is to be displayed}

\item{...}{other arguments passed to \code{str}}
}
\description{
Works like str() but displays units-relevant information more elegantly
}

