% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inspect.R
\name{verify_units}
\alias{verify_units}
\title{Assert that a unitted object has the expected units}
\usage{
verify_units(x, expected.units, return.values = list(x, NULL),
  violation.handler = stop, nounits.handler = warning)
}
\arguments{
\item{x}{character. The object to test for whether it has the expected units.}

\item{expected.units}{character. The expected units as a character [vector].}

\item{return.values}{list(2). The values to return if units are (1) as
expected, or (2) not as expected. Good options are list(x,NULL) and
list(TRUE, FALSE). Elements beyond position 2 are ignored.}

\item{violation.handler}{closure. The function to call if the expectation is
violated. Good options are stop, warning, message, and function(msg) {}.}
}
\value{
If units are as expected, returns return.values[[1]] (by default, x).
  Otherwise stops on an error (if violation.handler=stop) or returns
  return.values[[2]].
}
\description{
Throws an error (or calls a user-specified function) if the units are not as
expected. If the first element of return.values is x, verify_units(x) returns
x when units are as expected, and the function may therefore be chained with
other commands.
}

