% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.R
\docType{methods}
\name{write_unitted}
\alias{write_unitted}
\alias{write_unitted,unitted_data.frame-method}
\title{Write a unitted object to file.}
\usage{
write_unitted(x, ...)

\S4method{write_unitted}{unitted_data.frame}(x, comment.char = "#",
  sep = "\\t", row.names = FALSE, quote = FALSE, ...)
}
\arguments{
\item{x}{The unitted object to be written. Currently only implemented for
data.frames.}

\item{...}{other arguments passed to \code{\link{write.table}}}

\item{comment.char}{a single character, or "", with which to prepend the line
containing units information}

\item{sep}{the character string used to separate columns, as in
\code{\link{write.table}}}

\item{row.names}{logical. whether to write row names to the file, as in
\code{\link{write.table}}}

\item{quote}{logical. whether to place quotes around every data entry, as in
\code{\link{write.table}}}
}
\description{
Write a unitted object to file.

Write a unitted_data.frame to file.
}
\examples{
write_unitted(u(data.frame(x=1:3, y=9:7), c(x="grapes","seeds")))
write_unitted(u(data.frame(x=1:3, y=9:7), c(x="mg L^-1","ft^3 s^-1")), sep="\\t", file="practice.tsv")
file.remove("practice.tsv")
}

