% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{v}
\alias{deunitted}
\alias{deunitted,ANY-method}
\alias{deunitted,data.frame-method}
\alias{deunitted,list-method}
\alias{deunitted,unitted-method}
\alias{deunitted,unitted_data.frame-method}
\alias{deunitted,unitted_list-method}
\alias{v}
\title{Extract data from a unitted object}
\usage{
v(object, ...)

deunitted(object, ...)

\S4method{deunitted}{unitted}(object, ...)

\S4method{deunitted}{unitted_data.frame}(object, partial = FALSE, ...)

\S4method{deunitted}{unitted_list}(object, partial = FALSE, ...)

\S4method{deunitted}{data.frame}(object, ...)

\S4method{deunitted}{list}(object, ...)

\S4method{deunitted}{ANY}(object, ...)
}
\arguments{
\item{object}{A unitted object}

\item{...}{Other arguments passed to \code{deunitted}}

\item{partial}{logical. Should the data.frame or list be fully deunitted, such that
no elements will be left with units, or partially deunitted, such that the
object that is returned is not itself unitted but may have unitted
elements?}
}
\value{
A non-unitted data object
}
\description{
Unitted objects consist of data, units, and unitted class information. The
function \code{deunitted()} and its alias \code{v()} remove the units and
class information or, equivalently, extract the data from the unitted object.
}
\details{
In general, \code{deunitted()} and \code{v()} simply extract the
data part of the unitted object.

With unitted data.frames and lists, you have a choice between
  partial and complete deunitting (specified by the \code{partial} argument).
  Complete deunitting removes units and unitted class attributes both from
  the container (the data.frame or list) and its elements (the columns or
  list elements). Partial deunitting only removes the container units and
  unitted class attributes.

Non-unitted data.frames and lists may also be deunitted: This
  operation always removes units from the data.frame columns or list
  elements.

If \code{object} is neither a list nor a data.frame and is not
unitted, it will be returned intact.
}
\seealso{
\code{\link{u}()} and \code{\link{unitted}()} for the construction
  of unitted objects; \code{\linkS4class{unitted}} for the definition of the
  unitted class
}

