% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inspect.R
\docType{methods}
\name{get_units,unitted-method}
\alias{get_units,unitted-method}
\title{The exported, client-side method for acquiring an object's units in string representation.}
\usage{
\S4method{get_units}{unitted}(object, ...)
}
\arguments{
\item{y}{The object whose units should be returned}
}
\value{
A string representing one set of units
}
\description{
The exported, client-side method for acquiring an object's units in string representation.
}

