% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{get_units}
\alias{get_units}
\alias{get_units,ANY-method}
\alias{get_units,unitbundle-method}
\title{Gets the units as a string or strings.}
\usage{
get_units(object, ...)

\S4method{get_units}{ANY}(object, ...)

\S4method{get_units}{unitbundle}(object, delimiter = "|",
  rule = c("disambiguate", "never", "always"), ...)
}
\arguments{
\item{object}{The object from which to retrieve units}

\item{...}{Other arguments passed on to the type-specific implementations}

\item{delimiter}{A single-character string designating the delimiter that
should surround those units that are to be delimited according to
\code{rule}}

\item{rule}{character string indicating the rule by which each
unit within a unitdf will be delimited or not. The default, "disambiguate",
wraps only those units that contain spaces or "^" characters. "never" wraps
none, and "always" wraps all.}
}
\description{
Units can be acquired from objects of all types; types without units will
return NA.
}
\examples{
get_units(5)
get_units(unitbundle("king kong"))
}

