% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_join.R
\name{join}
\alias{join}
\alias{inner_join.unitted_data.frame}
\alias{left_join.unitted_data.frame}
\alias{right_join.unitted_data.frame}
\alias{full_join.unitted_data.frame}
\alias{semi_join.unitted_data.frame}
\alias{anti_join.unitted_data.frame}
\alias{inner_join.unitted_tbl_df}
\alias{left_join.unitted_tbl_df}
\alias{right_join.unitted_tbl_df}
\alias{full_join.unitted_tbl_df}
\alias{semi_join.unitted_tbl_df}
\alias{anti_join.unitted_tbl_df}
\title{Join two unitted data.frames}
\usage{
\method{inner_join}{unitted_data.frame}(x, y, by = NULL, copy = FALSE, ...)

\method{left_join}{unitted_data.frame}(x, y, by = NULL, copy = FALSE, ...)

\method{right_join}{unitted_data.frame}(x, y, by = NULL, copy = FALSE, ...)

\method{full_join}{unitted_data.frame}(x, y, by = NULL, copy = FALSE, ...)

\method{semi_join}{unitted_data.frame}(x, y, by = NULL, copy = FALSE, ...)

\method{anti_join}{unitted_data.frame}(x, y, by = NULL, copy = FALSE, ...)

\method{inner_join}{unitted_tbl_df}(x, y, by = NULL, copy = FALSE, ...)

\method{left_join}{unitted_tbl_df}(x, y, by = NULL, copy = FALSE, ...)

\method{right_join}{unitted_tbl_df}(x, y, by = NULL, copy = FALSE, ...)

\method{full_join}{unitted_tbl_df}(x, y, by = NULL, copy = FALSE, ...)

\method{semi_join}{unitted_tbl_df}(x, y, by = NULL, copy = FALSE, ...)

\method{anti_join}{unitted_tbl_df}(x, y, by = NULL, copy = FALSE, ...)
}
\arguments{
\item{x}{left-hand-side table to join}

\item{y}{right-hand-side table to join}

\item{by}{a character vector of variables to join by.  If \code{NULL}, the
  default, \code{join} will do a natural join, using all variables with
  common names across the two tables. A message lists the variables so
  that you can check they're right (to suppress the message, simply
  explicitly list the variables that you want to join).

  To join by different variables on x and y use a named vector.
  For example, \code{by = c("a" = "b")} will match \code{x.a} to
  \code{y.b}.}

\item{copy}{If \code{x} and \code{y} are not from the same data source,
and \code{copy} is \code{TRUE}, then \code{y} will be copied into the
same src as \code{x}.  This allows you to join tables across srcs, but
it is a potentially expensive operation so you must opt into it.}

\item{...}{other parameters passed onto methods}
}
\description{
Uses the corresponding dplyr join functions.
}
