% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-unitted.R
\docType{package}
\name{package-unitted}
\alias{package-unitted}
\alias{package-unitted-package}
\title{Rigorous unit tracking during data manipulation}
\description{
\code{unitted} helps you track the units of data values through all common
data manipulation activities - addition, multiplication, exponentiation,
subsetting, and so on.
}

