% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prepare_delimiter}
\alias{prepare_delimiter}
\title{Wrap delimiter in regexp escape characters if appropriate}
\usage{
prepare_delimiter(delim)
}
\arguments{
\item{delim}{character. A delimiter string to be wrapped in escape characters for use in regular expressions.}
}
\value{
The wrapped delimiter.
}
\description{
Helper to parse_units
}
\seealso{
Other unit-parsing functions: \code{\link{parse_units}};
  \code{\link{simplify_units}}; \code{\link{sort_units}}
}

