% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/display.R
\name{print.unitted}
\alias{print.unitted}
\title{Print a unitted object}
\usage{
\method{print}{unitted}(x, ...)
}
\arguments{
\item{x}{The unitted object}

\item{...}{other arguments to \code{\link{print}}}
}
\value{
An invisible unitted vector
}
\description{
For many data types, \code{print.unitted} includes the units in its output.
For data.frames, for example, it adds a second header row to display the
units of each column.
}

