% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\docType{methods}
\name{print.unitted}
\alias{print.unitted}
\alias{.unitted_print,data.frame-method}
\alias{.unitted_print,tbl_df-method}
\title{Print a unitted object}
\usage{
\method{print}{unitted}(x, ...)

\S4method{.unitted_print}{data.frame}(x, ..., digits = NULL, quote = FALSE,
  right = TRUE, row.names = TRUE)

\S4method{.unitted_print}{tbl_df}(x, ..., n = NULL, width = NULL,
  n_extra = NULL)
}
\arguments{
\item{x}{The unitted object}

\item{...}{other arguments to \code{\link{print}}}

\item{digits}{the minimum number of significant digits to be used: see
    \code{\link{print.default}}.}

\item{quote}{logical, indicating whether or not entries should be
    printed with surrounding quotes.}

\item{right}{logical, indicating whether or not strings should be
    right-aligned.  The default is right-alignment.}

\item{row.names}{logical (or character vector), indicating whether (or
    what) row names should be printed.}

\item{n}{As in tibble's \code{\link[tibble]{formatting}}, Number of rows to
show. If `NULL`, the default, will print all rows if less than option 
`tibble.print_max`. Otherwise, will print `tibble.print_min` rows.}

\item{width}{As in tibble's \code{\link[tibble]{formatting}}, Width of text
output to generate. This defaults to NULL, which means use 
`getOption("tibble.width")` or (if also NULL) `getOption("width")`; the 
latter displays only the columns that fit on one screen. You can also set 
`options(tibble.width = Inf)` to override this default and always print all
columns.}

\item{n_extra}{As in tibble's \code{\link[tibble]{formatting}}, Number of
extra columns to print abbreviated information for, if the width is too
small for the entire tibble. If `NULL`, the default, will print information
about at most `tibble.max_extra_cols` extra columns.}
}
\value{
An invisible unitted vector
}
\description{
For many data types, \code{print.unitted} includes the units in its output. 
For data.frames, for example, it adds a second header row to display the
units of each column.

Print a unitted data.frame

Print a unitted_tbl_df
}
\examples{
head(u(data.frame(x = u(1:500,"A"), y = u(runif(500),"B"), z = u(500:1,"C"))))
tibble::as_tibble(u(list(x = u(1:500,"A"), y = u(runif(500),"B"), z = u(500:1,"C"))))
}
