% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{select}
\alias{rename_.unitted_data.frame}
\alias{rename_.unitted_tbl_df}
\alias{select}
\alias{select_.unitted_data.frame}
\alias{select_.unitted_tbl_df}
\title{Column selection for unitted data.frames, tibbles, etc}
\usage{
\method{select_}{unitted_data.frame}(.data, ..., .dots)

\method{select_}{unitted_tbl_df}(.data, ..., .dots)

\method{rename_}{unitted_data.frame}(.data, ..., .dots)

\method{rename_}{unitted_tbl_df}(.data, ..., .dots)
}
\arguments{
\item{.data}{a unitted_data.frame}

\item{...}{standard dots, as in \code{\link[dplyr]{rename_}}}

\item{.dots}{nonstandard dots, as in \code{\link[dplyr]{rename_}}}
}
\value{
a unitted_data.frame after the \code{\link[dplyr]{rename_}} operation

a unitted_data.frame after the \code{\link[dplyr]{rename_}} operation
}
\description{
Column selection for unitted data.frames, tibbles, etc

Implements dplyr::select and dplyr::select_ for unitted_data.frames

Implements dplyr::select and dplyr::select_ for unitted_tbl_dfs

Implements dplyr::rename and dplyr::rename_ for unitted_data.frames

Implements dplyr::rename and dplyr::rename_ for unitted_tbl_dfs
}
\examples{
dplyr::select(u(data.frame(x=1:3, y=3:5, z=c("aa", "bb", "cc")), c("X","Y","Z")), a=y, x)
dplyr::select(tibble(u(data.frame(x=1:3, y=3:5, z=c("aa", "bb", "cc")), c("X","Y","Z"))), a=y, x)
df <- u(data.frame(x=1:3, y=3:5, z=c("aa", "bb", "cc")), c("X","Y","Z"))
dplyr::rename(df, a=y, beta=x)
df <- u(data.frame(x=1:3, y=3:5, z=c("aa", "bb", "cc")), c("X","Y","Z"))
dplyr::rename(tibble(df), a=y, beta=x)
}

