% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-unitbundle.R
\name{separate_units}
\alias{separate_units}
\title{Separate a unitbundle into a data frame of units and powers}
\usage{
separate_units(unitbdl)
}
\arguments{
\item{unitbdl}{a unitbundle to separate into units & powers}
}
\value{
a data.frame with columns for unit and power
}
\description{
Separate a unitbundle into a data frame of units and powers
}
