% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\docType{methods}
\name{show,unitted-method}
\alias{show,unitted-method}
\title{Show a unitted object}
\usage{
\S4method{show}{unitted}(object)
}
\arguments{
\item{object}{the unitted object to display}
}
\description{
Displays a unitted object, including its units and data type whenever possible
}
