% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-parse.R
\name{simplify_units}
\alias{simplify_units}
\title{Combine redundant units and eliminate units raised to the 0th power}
\usage{
simplify_units(unitdfs)
}
\arguments{
\item{unitdfs}{list of unit data.frames to be simplified}
}
\value{
list of simplified unit data.frames
}
\description{
Combine redundant units and eliminate units raised to the 0th power
}
\seealso{
Other unit-parsing functions: \code{\link{parse_units}},
  \code{\link{prepare_delimiter}}, \code{\link{sort_units}}
}
