% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{unitted_access}
\alias{unitted_access}
\alias{[.unitted}
\alias{[}
\alias{Extract}
\alias{[[.unitted}
\alias{[[}
\alias{$.unitted}
\alias{$}
\title{Extract parts of a unitted object}
\usage{
\method{[}{unitted}(x, ...)

\method{[[}{unitted}(x, ...)

\method{$}{unitted}(x, name)
}
\arguments{
\item{x}{The unitted data.frame, vector, etc. to be accessed}

\item{...}{Arguments passed to accessors, typically including one or more 
extraction indices}

\item{name}{The name of the column to be extracted}
}
\value{
A new data.frame, vector, etc. with the right units still attached.
}
\description{
The standard accessors \code{x[...]}, \code{x[[...]]}, and \code{x$name} are
all available for unitted objects.
}
\details{
\code{x[...]} extracts elements with their units.

\code{x[[...]]} drops most attributes but retains units.

\code{x$name} extracts a named column, with units, from a unitted data.frame
}
\seealso{
\code{\link{unitted_assign}} for assignment to parts of objects; 
  \code{\linkS4class{unitted}} for definition of the unitted class
}
