% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine.R
\name{unitted_bind}
\alias{bind}
\alias{cbind}
\alias{cbind.unitted}
\alias{rbind}
\alias{rbind.unitted}
\alias{unitted_bind}
\alias{unitted_cbind}
\alias{unitted_rbind}
\title{Bind unitted objects by row or column}
\usage{
rbind.unitted(...)

cbind.unitted(...)
}
\arguments{
\item{...}{unitted vectors, matrices, or data.frames}
}
\description{
Combines unitted objects as if they were not unitted, but enforces unit
consistency across the objects to be combined.
}
\details{
\code{deparse.level} is an integer controlling the construction of labels, as in
  the default rbind and cbind methods. Only deparse.level = 0 and 1 are
  available for unitted rbind/cbind calls.
}
\seealso{
Other unitted object manipulation: \code{\link{c,unitted-method}},
  \code{\link{c.unitted}}, \code{\link{unitted_c}};
  \code{\link{merge,ANY,unitted-method}},
  \code{\link{merge,unitted,ANY-method}},
  \code{\link{merge,unitted_data.frame,unitted_data.frame-method}},
  \code{\link{unitted_merge}}; \code{\link{rep.unitted}},
  \code{\link{rep.unitted_data.frame}},
  \code{\link{unitted_rep}}
}

