% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{unitted_math}
\alias{unitted_math}
\alias{Math.unitted}
\alias{Math}
\title{Group "Math" functions}
\usage{
\method{Math}{unitted}(x, ..., check.input.units = TRUE)
}
\arguments{
\item{x}{A vector (probably numeric or complex)}

\item{...}{Other arguments passed to the specific Math function}

\item{check.input.units}{logical. Should the units of x be checked for 
compatibility with the specific Math function? Functions abs, floor, 
ceiling, trunc, round, signif, and sqrt accept any units. Functions exp,
log, expm1, log1p, acos, asin, atan require that inputs are unitless.
Functions cos, sin, and tan require that inputs are "radians".}
}
\description{
Implements the S3 group generic, Math
}
