% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\docType{methods}
\name{unitted_merge}
\alias{merge,ANY,unitted-method}
\alias{merge,unitted,ANY-method}
\alias{merge,unitted_data.frame,unitted_data.frame-method}
\alias{unitted_merge}
\title{Merge unitted data.frames by one or more common columns}
\usage{
\S4method{merge}{unitted_data.frame,unitted_data.frame}(x, y, ...)

\S4method{merge}{unitted,ANY}(x, y, ...)

\S4method{merge}{ANY,unitted}(x, y, ...)
}
\arguments{
\item{x}{first unitted object to merge}

\item{y}{second unitted object to merge}

\item{...}{other arguments passed to \code{c}}
}
\description{
Merges unitted data.frames, ensuring units compatibility among the common
columns
}
\seealso{
\code{base::\link{merge}}

Other unitted object manipulation: \code{\link{unitted_bind}},
  \code{\link{unitted_c}}, \code{\link{unitted_rep}}
}

