% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arithmetic.R
\docType{methods}
\name{unitted_Ops}
\alias{unitted_Ops}
\alias{unitted_Ops,ANY-method}
\title{The workhorse method for unitted Ops.}
\usage{
unitted_Ops(.Generic, e1, e2)

\S4method{unitted_Ops}{ANY}(.Generic, e1, e2)
}
\arguments{
\item{.Generic}{A generic function name, as for Ops.unitted and the S4 generic Ops}

\item{e1}{An element, as for Ops.unitted and the S4 generic Ops}

\item{e2}{An element, as for Ops.unitted and the S4 generic Ops}
}
\description{
\code{unitted_Ops} works behind the scenes to ensure that operations on
unitted objects respect their units.
}

