% Generated by roxygen2 (4.0.1): do not edit by hand
\name{unitted_rep}
\alias{rep.unitted}
\alias{rep.unitted_data.frame}
\alias{unitted_rep}
\title{Replicate elements, maintaining the original units}
\usage{
\method{rep}{unitted}(x, ...)

\method{rep}{unitted_data.frame}(x, ...)
}
\description{
Wrapper for non-unitted rep() methods.
}
\seealso{
Other unitted object manipulation: \code{\link{bind}},
  \code{\link{cbind}}, \code{\link{cbind.unitted}},
  \code{\link{rbind}}, \code{\link{rbind.unitted}},
  \code{\link{unitted_bind}}, \code{\link{unitted_cbind}},
  \code{\link{unitted_rbind}}; \code{\link{c.unitted}},
  \code{\link{unitted_c}};
  \code{\link{merge,ANY,unitted-method}},
  \code{\link{merge,unitted,ANY-method}},
  \code{\link{unitted_merge}}
}

